Split history and logbook
